Week 2 compleet

+++ Betaalwijze.java
@@ -2,13 +2,13 @@ import java.math.*;
 
 public abstract class Betaalwijze {
 
-    protected BigDecimal saldo;
+    protected double saldo;
 
     /**
      * Methode om krediet te initialiseren
      * @param saldo
      */
-    public void setSaldo(BigDecimal saldo) {
+    public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
 
@@ -18,5 +18,5 @@ public abstract class Betaalwijze {
      * @param tebetalen
      * @return Boolean om te kijken of er voldoende saldo is
      */
-    public abstract boolean betaal(BigDecimal tebetalen);
+    public abstract boolean betaal(double tebetalen);
 }
