+++ Contant.java
@@ -4,7 +4,7 @@ public class Contant extends Betaalwijze {
     /**
      * Methode om betaling af te handelen
      */
-    public boolean betaal(BigDecimal tebetalen) {
+    public boolean betaal(double tebetalen) {
         // method body omitted
         return true; // TEMPORARY
     }
