+++ Pinpas.java
@@ -2,20 +2,20 @@ import java.math.*;
 
 public class Pinpas extends Betaalwijze {
 
-    private BigDecimal kredietlimiet;
+    private double kredietlimiet;
 
     /**
      * Methode om kredietlimiet te zetten
      * @param kredietlimiet
      */
-    public void setKredietLimiet(BigDecimal kredietlimiet) {
+    public void setKredietLimiet(double kredietlimiet) {
         // method body omitted
     }
 
     /**
      * Methode om betaling af te handelen
      */
-    public boolean betaal(BigDecimal tebetalen) {
+    public boolean betaal(double tebetalen) {
         // method body omitted
         return true; // TEMPORARY
     }
